server/requirements.txt
fastapi>=0.115.2
uvicorn>=0.35.0
pydantic>=2.7.0
openai==2.30.0
requests==2.32.3
python-multipart==0.0.20
openenv-core[core]==0.2.3
python-dotenv==1.2.2


